mobile-app/backend/requirements.txt
Flask
flask-cors
psycopg2-binary
python-dotenv
bcrypt
PyJWT
qrcode
Pillow